+++ examples/tutorial_oss.py
@@ -4,10 +4,9 @@ from typing import Optional, Union, cast
 import torch
 import torch.distributed as dist
 import torch.multiprocessing as mp
-import torch.nn as nn
-import torch.nn.functional as F
 
 from fairscale.optim.oss import OSS
+from helpers import dist_init, getModel, getData, getLossFun
 
 WORLD_SIZE = 2
 EPOCHS = 3
@@ -15,34 +14,15 @@ EPOCHS = 3
 DEVICE = "cuda" if torch.cuda.is_available() else "cpu"
 
 
-def dist_init(rank, world_size):
-    backend = dist.Backend.NCCL if torch.cuda.is_available() else dist.Backend.GLOO  # type: ignore
-    print(f"Using backend: {backend}")
-    dist.init_process_group(backend=backend, init_method="tcp://localhost:29501", rank=rank, world_size=world_size)
-
-
-def getModel():
-    return nn.Sequential(torch.nn.Linear(10, 10), torch.nn.ReLU(), torch.nn.Linear(10, 5))
-
-
-def getData():
-    target = torch.randint(0, 2, size=(20, 1)).squeeze()
-    data = torch.randn(20, 10)
-    return [(data, target)]
-
-
-def getLossFun():
-    return F.nll_loss
-
-
 def train(rank: int, world_size: int, epochs: int, use_oss: bool):
 
     # DDP
     dist_init(rank, world_size)
+    rank = torch.device("cpu") if DEVICE == "cpu" else rank
 
     # Problem statement
     model = getModel().to(rank)
-    dataloader = getData()
+    dataloader = getData(n_batches=1)
     loss_fn = getLossFun()
 
     optimizer: Optional[Union[OSS, torch.optim.SGD]] = None
@@ -52,7 +32,7 @@ def train(rank: int, world_size: int, epochs: int, use_oss: bool):
     else:
         base_optimizer = torch.optim.SGD
         base_optimizer_arguments = {"lr": 1e-4}  # any optimizer specific arguments, LR, momentum, etc...
-        optimizer = OSS(params=model.parameters(), optim=base_optimizer, default=base_optimizer_arguments)
+        optimizer = OSS(params=model.parameters(), optim=base_optimizer, **base_optimizer_arguments)
 
     training_start = time.monotonic()
     # Any relevant training loop, nothing specific to OSS. For example:
@@ -82,10 +62,11 @@ def train(rank: int, world_size: int, epochs: int, use_oss: bool):
             print(f"Loss: {loss.item()}")
 
     training_end = time.monotonic()
-    max_memory = torch.cuda.max_memory_allocated(rank)
-
     print(f"[{dist.get_rank()}] : Training done. {training_end-training_start:.2f} sec")
-    print(f"[{dist.get_rank()}] : Peak memory {max_memory:.1f}MiB")
+
+    if DEVICE == "cuda":
+        max_memory = torch.cuda.max_memory_allocated(rank)
+        print(f"[{dist.get_rank()}] : Peak memory {max_memory:.1f}MiB")
 
 
 if __name__ == "__main__":
