+++ examples/tutorial_pipe_rpc.py
@@ -4,13 +4,12 @@
 import os
 
 import torch
-import torch.nn as nn
-import torch.nn.functional as F
 import torch.optim as optim
 import torch_pg
 
 import fairscale
 from fairscale.nn.model_parallel import initialize_model_parallel
+from helpers import dist_init, getModel, getData, getLossFun
 
 
 def register_optimizer(ctx, model):
@@ -28,7 +27,7 @@ def run(rank, world_size):
     torch_pg.init_mpi()
     os.environ["MASTER_ADDR"] = "localhost"
     os.environ["MASTER_PORT"] = "10638"
-    torch.distributed.init_process_group(backend="nccl", rank=rank, world_size=world_size)
+    dist_init(rank, world_size) # FIXME (supports gloo)
     os.environ["MASTER_PORT"] = "10639"
     torch.distributed.rpc.init_rpc(f"worker{rank}", rank=rank, world_size=world_size)
     initialize_model_parallel(1, world_size, pipeline_backend="mpi")
@@ -38,10 +37,9 @@ def run(rank, world_size):
         torch.distributed.rpc.shutdown()
         return
 
-    model = nn.Sequential(torch.nn.Linear(10, 10), torch.nn.ReLU(), torch.nn.Linear(10, 5))
-    target = torch.randint(0, 2, size=(20, 1)).squeeze()
-    data = torch.randn(20, 10)
-    loss_fn = F.nll_loss
+    model = getModel()
+    data, target = getData()[0]
+    loss_fn = getLossFun()
 
     device = torch.device("cuda", rank)
